Ct/v2.0

+++ apps/server/cypress/support/e2e.ts
@@ -92,13 +92,15 @@ beforeEach(() => {
 });
 
 afterEach(() => {
-  const widgetDemoAccessToken = Cypress.env(
-    WIDGET_DEMO_DELETE_USER_ACCESS_TOKEN_ENV,
-  );
+  if (Cypress.env("delete_mx_user_after_test") == true) {
+    const widgetDemoAccessToken = Cypress.env(
+      WIDGET_DEMO_DELETE_USER_ACCESS_TOKEN_ENV,
+    );
 
-  deleteMxUser({
-    headers: {
-      authorization: `Bearer ${widgetDemoAccessToken}`,
-    },
-  });
+    deleteMxUser({
+      headers: {
+        authorization: `Bearer ${widgetDemoAccessToken}`,
+      },
+    });
+  }
 });
